Created LICENSE
Added MIT License

This commit adds the MIT License to the project's repository. The MIT License is an open-source license that grants permission for others to use, modify, and distribute the project's code while including the original copyright notice and disclaiming liability. This license promotes collaboration and ensures that the project remains open and free for anyone to use.

By adding this license, we make it clear how others can use and contribute to the project while protecting the project's integrity and providing proper attribution.

Please refer to the LICENSE file for the full text of the MIT License.